Move AbstractTestSupervisor
