+++ mephisto/data_model/test/utils.py
@@ -3,8 +3,16 @@
 # Copyright (c) Facebook, Inc. and its affiliates.
 # This source code is licensed under the MIT license found in the
 # LICENSE file in the root directory of this source tree.
+import os
+import shutil
+import tempfile
+from abc import ABC
 
-from typing import Optional, Tuple
+from typing import Optional, Tuple, ClassVar, Type, List
+
+from mephisto.core.local_database import LocalMephistoDB
+from mephisto.core.task_launcher import TaskLauncher
+from mephisto.data_model.blueprint import SharedTaskState
 
 from mephisto.data_model.database import (
     MephistoDB,
@@ -14,18 +22,19 @@ from mephisto.data_model.database import (
 )
 
 from mephisto.data_model.agent import Agent
-from mephisto.data_model.assignment import Unit, Assignment
+from mephisto.data_model.assignment import Unit, Assignment, InitializationData
 from mephisto.data_model.task_config import TaskConfig
 from mephisto.data_model.requester import Requester
 from mephisto.data_model.task import Task, TaskRun
 from omegaconf import OmegaConf
 import json
 
-from mephisto.providers.mock.mock_provider import MockProviderArgs
+from mephisto.providers.mock.mock_provider import MockProviderArgs, MockProvider
 from mephisto.server.blueprints.mock.mock_blueprint import MockBlueprintArgs
-from mephisto.server.architects.mock_architect import MockArchitectArgs
+from mephisto.server.architects.mock_architect import MockArchitectArgs, MockArchitect
 from mephisto.data_model.task_config import TaskConfigArgs
 from mephisto.core.hydra_config import MephistoConfig
+from mephisto.server.blueprints.mock.mock_task_runner import MockTaskRunner
 
 MOCK_TASK_ARGS = TaskConfigArgs(
     task_title="title",
@@ -178,3 +187,60 @@ def make_completed_unit(db: MephistoDB) -> str:
     unit = Unit(db, unit_id)
     unit.sync_status()
     return unit.db_id
+
+
+EMPTY_STATE = SharedTaskState()
+
+
+class AbstractTestSupervisor(ABC):
+    """
+    Abstract test class utilizing the Mephisto Supervisor.
+
+    Subclasses will use the MockArchitect and the MockProvider, and they must specify a
+    BlueprintClass to test.
+    """
+
+    BlueprintClass: ClassVar[Type["Blueprint"]]
+
+    def setUp(self):
+        self.data_dir = tempfile.mkdtemp()
+        database_path = os.path.join(self.data_dir, "mephisto.db")
+        self.db = LocalMephistoDB(database_path)
+        self.task_id = self.db.new_task("test_mock", self.BlueprintClass.BLUEPRINT_TYPE)
+        self.task_run_id = get_test_task_run(self.db)
+        self.task_run = TaskRun(self.db, self.task_run_id)
+
+        architect_config = OmegaConf.structured(
+            MephistoConfig(architect=MockArchitectArgs(should_run_server=True))
+        )
+
+        self.architect = MockArchitect(
+            self.db, architect_config, EMPTY_STATE, self.task_run, self.data_dir
+        )
+        self.architect.prepare()
+        self.architect.deploy()
+        self.urls = self.architect._get_socket_urls()  # FIXME
+        self.url = self.urls[0]
+        self.provider = MockProvider(self.db)
+        self.provider.setup_resources_for_task_run(
+            self.task_run, self.task_run.args, EMPTY_STATE, self.url
+        )
+        self.launcher = TaskLauncher(
+            self.db, self.task_run, self.get_mock_assignment_data_array()
+        )
+        self.launcher.create_assignments()
+        self.launcher.launch_units(self.url)
+        self.sup = None
+
+    def tearDown(self):
+        if self.sup is not None:
+            self.sup.shutdown()
+        self.launcher.expire_units()
+        self.architect.cleanup()
+        self.architect.shutdown()
+        self.db.shutdown()
+        shutil.rmtree(self.data_dir, ignore_errors=True)
+
+    def get_mock_assignment_data_array(self) -> List[InitializationData]:
+        mock_data = MockTaskRunner.get_mock_assignment_data()
+        return [mock_data, mock_data]
